refactor storyboard related code to share common utlities
