+++ server/core/lib/job-queue/handlers/generate-storyboard.ts
@@ -15,9 +15,9 @@ import { VideoPathManager } from '@server/lib/video-path-manager.js'
 import { getImageSizeFromWorker } from '@server/lib/worker/parent-process.js'
 import { StoryboardModel } from '@server/models/video/storyboard.js'
 import { VideoModel } from '@server/models/video/video.js'
-import { MVideo } from '@server/types/models/index.js'
 import { Job } from 'bullmq'
 import { join } from 'path'
+import { buildSpritesMetadata, findGridSize } from './shared/storyboard-utils.js'
 
 const lTagsBase = loggerTagsFactory('storyboard')
 
@@ -153,35 +153,3 @@ export {
   processGenerateStoryboard
 }
 
-function buildSpritesMetadata (options: {
-  video: MVideo
-}) {
-  const { video } = options
-
-  if (video.duration < 3) return { spriteDuration: undefined, totalSprites: 0 }
-
-  const maxSprites = Math.min(Math.ceil(video.duration), STORYBOARD.SPRITES_MAX_EDGE_COUNT * STORYBOARD.SPRITES_MAX_EDGE_COUNT)
-
-  const spriteDuration = Math.ceil(video.duration / maxSprites)
-  const totalSprites = Math.ceil(video.duration / spriteDuration)
-
-  // We can generate a single line so we don't need a prime number
-  if (totalSprites <= STORYBOARD.SPRITES_MAX_EDGE_COUNT) return { spriteDuration, totalSprites }
-
-  return { spriteDuration, totalSprites }
-}
-
-function findGridSize (options: {
-  toFind: number
-  maxEdgeCount: number
-}) {
-  const { toFind, maxEdgeCount } = options
-
-  for (let i = 1; i <= maxEdgeCount; i++) {
-    for (let j = i; j <= maxEdgeCount; j++) {
-      if (toFind <= i * j) return { width: j, height: i }
-    }
-  }
-
-  throw new Error(`Could not find grid size (to find: ${toFind}, max edge count: ${maxEdgeCount}`)
-}
