+++ server/core/lib/runners/job-handlers/video-storyboard-job-handler.ts
@@ -23,6 +23,7 @@ import { sequelizeTypescript } from '@server/initializers/database.js'
 import { federateVideoIfNeeded } from '@server/lib/activitypub/videos/index.js'
 import { generateImageFilename } from '@server/helpers/image-utils.js'
 import { getImageSizeFromWorker } from '@server/lib/worker/parent-process.js'
+import { buildSpritesMetadataFromDuration, findGridSize } from '../../job-queue/handlers/shared/storyboard-utils.js'
 
 type CreateOptions = {
   videoUUID: string
@@ -58,7 +59,7 @@ export class VideoStoryboardJobHandler extends AbstractJobHandler<CreateOptions,
       }
     })
 
-    const { spriteDuration, totalSprites } = buildSpritesMetadata({ duration })
+    const { spriteDuration, totalSprites } = buildSpritesMetadataFromDuration({ duration })
     const spritesCount = findGridSize({ toFind: totalSprites, maxEdgeCount: STORYBOARD.SPRITES_MAX_EDGE_COUNT })
 
     const payload: RunnerJobGenerateStoryboardPayload = {
@@ -139,29 +140,3 @@ export class VideoStoryboardJobHandler extends AbstractJobHandler<CreateOptions,
   protected specificError (_options: { runnerJob: MRunnerJob }) {}
   protected specificCancel (_options: { runnerJob: MRunnerJob }) {}
 }
-
-function buildSpritesMetadata (options: { duration: number }) {
-  const { duration } = options
-  if (duration < 3) return { spriteDuration: undefined, totalSprites: 0 }
-
-  const maxSprites = Math.min(Math.ceil(duration), STORYBOARD.SPRITES_MAX_EDGE_COUNT * STORYBOARD.SPRITES_MAX_EDGE_COUNT)
-
-  const spriteDuration = Math.ceil(duration / maxSprites)
-  const totalSprites = Math.ceil(duration / spriteDuration)
-
-  if (totalSprites <= STORYBOARD.SPRITES_MAX_EDGE_COUNT) return { spriteDuration, totalSprites }
-
-  return { spriteDuration, totalSprites }
-}
-
-function findGridSize (options: { toFind: number, maxEdgeCount: number }) {
-  const { toFind, maxEdgeCount } = options
-  for (let i = 1; i <= maxEdgeCount; i++) {
-    for (let j = i; j <= maxEdgeCount; j++) {
-      if (toFind <= i * j) return { width: j, height: i }
-    }
-  }
-  throw new Error(`Could not find grid size (to find: ${toFind}, max edge count: ${maxEdgeCount}`)
-}
-
-
